+++ modules/nlu-extras/src/config.ts
@@ -4,7 +4,7 @@ export interface Config {
   /**
    * Specify the primary NLU engine that will be used by Botpress.
    *
-   * NOTE: Botpress NLU always run and can't be disabled. If an other NLU is specified as primary,
+   * NOTE: Botpress NLU always run and can't be disabled. If another NLU is specified as primary,
    * it will run after Botpress and will overwrite the prediction results.
    * @default 'botpress-nlu'
    */